# Vectorize `SpinOp.simplify` and `VibrationalOp.simplify`