@woodsp-ibm 2022-04-01
This is all about performance right. You had some figures in the prior PR, do you have any timings for the before/after in this case.