@t-imamichi 2022-04-06
I made a microbenchmark.

```python
import random
from timeit import timeit

from qiskit_nature.operators.second_quantization import VibrationalOp, SpinOp


def vb_random_op(k, n):
    random.seed(123)
    op = sum(
        (VibrationalOp("".join(random.choices("+-ENI", k=k)), k, 1) * random.random())
        for _ in range(n))
    return op


def spin_random_op(k, n):
    random.seed(123)
    op = sum(
        (SpinOp("".join(random.choices("IXYZ", k=k))) * random.random())
        for _ in range(n))
    return op


def bench_simplify(k, n):
    op = vb_random_op(k, n)
    print(f"vibrational op, {k} {n}: {timeit(lambda: op.simplify(), number=100)} sec")

    op = spin_random_op(k, n)
    print(f"spin op, {k} {n}: {timeit(lambda: op.simplify(), number=100)} sec")


bench_simplify(10, 1000)
```

main
```
vibrational op, 10 1000: 3.0248702310000226 sec
spin op, 10 1000: 3.114273889999822 sec
```

this PR
```
vibrational op, 10 1000: 0.16983310300020094 sec
spin op, 10 1000: 0.19599379700048303 sec
```