@kevinsung 2022-04-06
I wonder if we should change these to a dict-based implementation, as is done in `FermionicOp`. That would have better asymptotic complexity (O(n) instead of O(n log n)). It might be slower for small operators, but if that's an issue it seems we can port it to Rust as was done at https://github.com/Qiskit/qiskit-terra/pull/7656. In any case, that should probably be done separately after merging this PR.