@t-imamichi 2022-04-06
If `simplify` is still a bottleneck of some applications even with this PR, it would be good to think of the rust version. I hope we add more benchmarking codes to https://github.com/Qiskit/qiskit-app-benchmarks. Currently, there is only protein folding in nature.